+++ crates/adapters/betfair/examples/betfair_backtest.rs
@@ -153,7 +153,7 @@ fn main() -> anyhow::Result<()> {
             .account_type(AccountType::Cash)
             .book_type(BookType::L2_MBP)
             .starting_balances(vec![Money::from("1_000_000 GBP")])
-            .build(),
+            .build()?,
     )?;
 
     for instrument in instruments.values() {